Move utility method to a different class
